Change allocation buffer for article tree.
Previous impelemntation used std::deque as allocation buffer for articles and nodes in tree.  This was not as efficient as it should have been since the gcc implementation only allocated 512 bytes at a time.  I've added a paired down version optimized for this usage that allocates 8KB at a time.