Attempt to fix display of threads with unread articles.
When using the "Show Matching Articles' Threads" and "Match Only Unread
Articles" options together, some articles may not be shown at all. In
particular, when the first article in a particular thread either doesn't exist
or has never been retrieved, Pan ignores that thread entirely in the display.

The problem seems to stem from accumulate_descendants, which will only traverse
the tree if the node has an associated article. In many cases, such as above,
the node doesn't have an article, but it does have child nodes which do have
articles.

This patch fixes the problem for me, and it has never caused a crash that I am
aware of, but I am insufficiently familiar with Pan's internals to definitely
proclaim this the best solution.

Signed-off-by: Jason Lynch <jason@calindora.com>